refactor: Refactor custom login page
Modify the 'auth/signin' server component.
In this component, retrieve the provider using the 'getServerSession' method and pass the provider as a prop to the SignIn client component.
Create a new file called 'SignIn.tsx' in the 'components' folder.
The SignIn component handles the UI and sign-in with Google using the next-auth provider.